refactor(db): apply shared columns review
