+++ packages/db/src/domains/ship/schema.ts
@@ -1,6 +1,6 @@
 import { integer, pgTable, serial, varchar } from 'drizzle-orm/pg-core';
 
-import { buildTimestampColumns } from '../../shared/columns/timestamp-columns.js';
+import { buildTimestampColumns } from '../../shared/columns/index.js';
 import { player } from '../player/schema.js';
 
 export const ship = pgTable('ship', {
