+++ packages/db/src/shared/columns/pokemon-types-column.ts
@@ -1,10 +1,12 @@
 import { sql } from 'drizzle-orm';
 
-import { pokemonType } from '../enums/pokemon-type.js';
+import { pokemonType } from '../enums/pokemon-type-enum.js';
 
 export function buildPokemonTypesColumn() {
-  return pokemonType('types')
-    .array()
-    .notNull()
-    .default(sql`'{}'::pokemon_type[]`);
+  return {
+    types: pokemonType('types')
+      .array()
+      .notNull()
+      .default(sql`'{}'::pokemon_type[]`),
+  };
 }
