+++ packages/db/src/shared/enums/pokemon-type-enum.ts
