+++ packages/db/src/shared/enums/rarity-enum.ts
